add new test
